Added dtype kwarg to matrix.copy and changed det_inv to use dtype=float to prevent incorrect det calculations with matricies of dtype=int.

+++ ulinalg.py
@@ -58,7 +58,7 @@ def det_inv(x):
     else:
         # divide each row element by [0] to give a one in the first position
         # (may have to find a row to switch with if first element is 0)
-        x = x.copy()
+        x = x.copy(dtype=float)
         inverse = eye(len(x), dtype=float)
         sign = 1
         factors = []
